Move things to their (maybe?) final locations
